Moved and implemented query selection when moving to account_update
Changes: ALL html is moved into the ROOT folder of public
Reason: You dont need an html folder to host all of it. public serves that purpose. You also dont need to .use ALL of the paths. Currently we will just have app.use("/", express.static('./public')); to serve as the root folder. This simplifies all of the .uses

Changes: added cors using all cors because fetch is weird and the browser wouldnt let me move pages without it.

Changes: all paths now reflect the new root folder changes. so if you want an html file you just do /public/*htmlfile* if you are in the backend server or /*htmlfile* if you are in the frontend

comments: still working on the admin_edit page because of using query selectors and such.